Check before uninstalling cygwin.
Former-commit-id: 793addcdf565c1684fbfc94126030ab422308fd8
Former-commit-id: 65d58160bd724e80786b081a665612a2b91d89fc

+++ SetupCygwinInstallAction/CygwinInstaller.cs
@@ -140,11 +140,16 @@ namespace SetupCygwinInstallAction {
 
             try {
                 System.IO.DirectoryInfo di = new DirectoryInfo(LOCALAPPDATA);
-
+                
                 foreach (FileInfo file in di.GetFiles()) {
                     file.Delete();
                 }
                 foreach (DirectoryInfo dir in di.GetDirectories()) {
+                    if(dir.FullName == CYGWIN_LOC) {
+                        if( System.Windows.Forms.MessageBox.Show("Local Astrometry.Net Client detected. Uninstall?", "", System.Windows.Forms.MessageBoxButtons.YesNo) != System.Windows.Forms.DialogResult.Yes) {
+                            continue;
+                        }
+                    }
                     dir.Delete(true);
                 }
                 di.Delete();
